Add address badge to table of contents
By adding the address badge to the table of contents, it mirrors the official datasheet where they had the project named followed by the address in square brackets.

This is achieved by first creating a fake heading which contains the badge, and then hiding it. Then, the real heading is created and moved up into the empty space that is now left by the hidden heading.