Fix overflowing project titles
This bug was caused by having to have two different titles for one project, just so that the table of contents could show the project's address. Therefore, there was a need to somehow calculate the distance that the real title should be moved up by, so that everything looks normal.

The real distance is calculated by measuring the height of the bock that the fake title is in. This block has a max width of `page.width - 2.5cm` at the moment, but this might have to be tweaked more. By limiting the block width, it ensures that the title properly overflows to the next line, and that the actual height reflects that.

Fixes #1