Add colourful splash chapter page & `projects` parameter to datasheet
This commit adds a callable function to generate a full-page colour chapter page that only contains the title of the chapter and the footer.

You can pass a boolean value to `invert-text-colour` to make sure that the contrast is kept proper, but that is not something that is automated at the moment.

The `datasheet()` function now also takes a `projects` parameter, which can contain the content of all the projects in the datasheet -- this is really intended for passing `include projects.typ` to it from the main file, but you should be able to use regular content with it too. The file path cannot be simply passed to it because otherwise the compiler will not see the files properly.

Fixes #3