Add parameter to select which chip pinout table to render
This is intended to help support multiple shuttles, as the harness used changes. The currently defined `new_frame` pinout first appears in IHP 0p2, and `caravel` first appears in TT06.

I'm not sure what the technical/accurate/canonical names for these pinouts are, so they might change in the future, but this is what I'm going with at the moment.

Ref #8