Add chapter title to footer
This commit fixes some formatting issues with the footer, namely the table of contents title being way too large and thus not being able to be used. This was caused by passing `content` to `outline()` that contained formatting, so now it only contains text. The styling for H1 was updated to add some space below the title.

Ref #7