Fix alternating footer content
When the page counter was reset because we were transitioning from the table of contents into the real content, it would mess up the ordering because the function itself was dependent on `counter(page)`. So when finishing on an odd page in the table of contents, and then resetting the counter to 1, it would mean that the footer content would be permanently flipped.

To resolve this, a stateful variable was added (`_flip_footer_ordering`), that has its state updated after the table of contents to determine whether the footer ordering needs to be flipped.

At the moment this solution feels hacky, so reopen #7