Update title formatting for splash chapter pages
For some reason show rules aren't working for headings in here, but even if they did, they would mess up the footer.

For this reason, we'll just make big text and hide the regular heading, in the hopes that if you jump to it, it'll look like the big title.