Fix PDF outline rendering
Due to the way that the address badge was implemented, the PDF outline would render `---0 Chip ROM`, instead of just having the title of the project - the `---0` was unwanted.

By alternating which heading is used for the outline and which is used for the bookmark, we can make sure that the PDF outline looks good without affecting the content of the datasheet.

Closes #25

I think by doing this alternating outlined/bookmarked title, it also fixed the issue where the project content was at the wrong heading level.

Closes #26