Add QR and optional sponsor info
This info + QR code will be placed on the splash chapter page of the "Chip Renders" section.

This commit updates the function signature for `splash_chapter_page()`, allowing it to accept additional content that will be placed in a white box at the bottom of the chapter title page. This `additional-content` parameter is used when making the "Chip Renders" chapter to pass in a generated QR code and sponsor logo/text.

Closes #24