Fix overflowing project titles
For real this time, I hope.

Instead of messing around trying to measure the height of the fake title and then trying to move the real one up by that amount, we can instead use `place()` to make it so that the fake title doesn't affect the layout, even though we're also using `hide()`.

By doing that , we eliminate that problem entirely and are free to place the real heading down like normal.

Closes #1