Share styling between datasheet and project
This should make it so that you're able to preview either the datasheet or a project, and have the same styling apply.

It does ommit some things though, like coloured links and footer content. I think this will be good enough to show to the end user once they've run their docs build.

Close #12